refactor(routes/access): replace the previous auth route with access route
